requirements.txt
web3==7.10.0
python-dotenv==1.0.1
requests==2.31.0
tabulate>=0.8.10
python-telegram-bot==22.0